+++ CHANGELOG.md
@@ -5,3 +5,10 @@
 ### Changed
 
 - Initial Release
+
+## v1.1.0
+
+### Changed
+
+- Add ima.py as related testcase
+- Minor code refactor for report.py by moving the get_file_type and create_iterator to class Utils